6-Front End with Streamlit/requirements(1).txt
streamlit==0.84.0
requests==2.25.1
st-annotated-text==2.0.0
